% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbes.R
\name{extractBV}
\alias{extractBV}
\title{Extracts measurements from a BV table}
\usage{
extractBV(BVtable, BVtypes, dataTypes)
}
\arguments{
\item{BVtable}{data.table() with column names following the RDBES (v 1.17) R-Name specification}

\item{BVtypes}{a character() vector identifying the BVtypes to introduce as columns (other record types are discarded)}

\item{dataTypes}{The desired data types for the columns. Currently supported: integer, numeric, character and logical}
}
\value{
data.table()
}
\description{
Reshapes BV table into a data.table with each row corresponding to one measured specimen,
 and the different BV types introduced as individual columns (e.g. age, length weight).
 @details
 Checks and halts for heterogenity in length units, weight untits, length measurements, or BVstratification
}
\examples{
extractBV(NORportsampling2018$BV, c("Age", "Length", "Weight"), c("integer", "numeric", "numeric"))
}
