% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbes.R
\name{plotSAnas}
\alias{plotSAnas}
\title{Missing data histograms}
\usage{
plotSAnas(flatRDBES, var)
}
\arguments{
\item{flatRDBES}{flatRDBES data.table() with column names following the RDBES (v 1.17) R-Name specification.}

\item{var}{variable to plot missing data for (column in flatRDBES)}
}
\description{
Plots the number and fraction of missing data pr sample (SAid)
}
\examples{
 ages <- extractBV(NORportsampling2018$BV, c("Age"), c("integer"))
 agesamples <- merge(ages, NORportsampling2018$SA, by="SAid")
 plotSAnas(agesamples, c("Age"))
}
