% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recaPrepExample-datadoc.R
\docType{data}
\name{recaPrepExample}
\alias{recaPrepExample}
\title{Example output from \code{\link[prepRECA]{prepRECA}}}
\format{An object of class \code{list} of length 5.}
\usage{
recaPrepExample
}
\description{
Example output from \code{\link[prepRECA]{prepRECA}}
}
\keyword{datasets}
