% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recaRunExample-datadoc.R
\docType{data}
\name{recaRunExample}
\alias{recaRunExample}
\title{Example output from \code{\link[prepRECA]{runRECA}}}
\format{An object of class \code{list} of length 3.}
\usage{
recaRunExample
}
\description{
Example output from \code{\link[prepRECA]{runRECA}}
}
\keyword{datasets}
