% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbes.R
\name{warningsDataCompleteness}
\alias{warningsDataCompleteness}
\title{Produce warnings for incomplete data}
\usage{
warningsDataCompleteness(datatable, variables)
}
\arguments{
\item{datatable}{data.table() to be checked for completeness}

\item{variables}{character() vector with variables (column names of datatable) to check for completeness}
}
\description{
Produce warnings for columns with NA values, listing the number of NA values.
}
