% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbes.R
\name{warningsRecaApplicability}
\alias{warningsRecaApplicability}
\title{Produce warnings for R-ECA}
\usage{
warningsRecaApplicability(flatRDBES)
}
\arguments{
\item{flatRDBES}{data.table() with column names following the RDBES (v 1.17) R-Name specification.}
}
\description{
Checks for common concerns when configuring R-ECA
}
\details{
Checks for heterogenity in fish presentation
 Checks for stratification
 Checks for multiple sampling programs
}
\examples{
warningsRecaApplicability(NORportsampling2018$SA)
}
